deletes deprecated histograms.js file, moves Heatmap.svelte
